+++ cmd/serverversion.go
@@ -3,7 +3,6 @@ package cmd
 import (
 	"github.com/spf13/cobra"
 	"log"
-	sonarFactory "sonarci/sonar/factory"
 )
 
 func NewServerVersionCmd() *cobra.Command {
@@ -25,7 +24,7 @@ func serverVersion(cmd *cobra.Command, args []string) {
 		return
 	}
 
-	api := sonarFactory.CreateLatestSonarRestApi(pFlags.Server, pFlags.Token, pFlags.Timeout)
+	api := createSonarApi(pFlags.Server, pFlags.Token, pFlags.Timeout)
 	version, err := api.GetServerVersion()
 	if err != nil {
 		log.Fatal("Failure to get server version: ", err)
