+++ cmd/validate.go
@@ -3,7 +3,6 @@ package cmd
 import (
 	"github.com/spf13/cobra"
 	"log"
-	sonarFactory "sonarci/sonar/factory"
 )
 
 const (
@@ -68,7 +67,7 @@ func validateBranch(cmd *cobra.Command, args []string) {
 		return
 	}
 
-	api := sonarFactory.CreateLatestSonarRestApi(pFlags.Server, pFlags.Token, pFlags.Timeout)
+	api := createSonarApi(pFlags.Server, pFlags.Token, pFlags.Timeout)
 	err := api.ValidateBranch(project, branch)
 	if err != nil {
 		log.Fatal(err)
@@ -88,7 +87,7 @@ func validatePullRequest(cmd *cobra.Command, args []string) {
 		return
 	}
 
-	api := sonarFactory.CreateLatestSonarRestApi(pFlags.Server, pFlags.Token, pFlags.Timeout)
+	api := createSonarApi(pFlags.Server, pFlags.Token, pFlags.Timeout)
 	err := api.ValidatePullRequest(project, pr)
 	if err != nil {
 		log.Fatal(err)
