+++ net/connection.go
@@ -1,5 +1,6 @@
 package net
 
 type Connection interface {
+	GetHostServer() string
 	DoGet(route string) (<-chan []byte, <-chan error)
 }
