+++ net/factory/factory.go
@@ -6,6 +6,6 @@ import (
 	"time"
 )
 
-func CreateHttpConnection(server string, token string, timeout time.Duration) net.Connection {
+func CreateConnection(server string, token string, timeout time.Duration) net.Connection {
 	return http.NewConnection(server, token, timeout)
 }
