+++ net/http/http.go
@@ -13,13 +13,17 @@ import (
 )
 
 type Connection struct {
-	Server  string
-	Token   string
-	Timeout time.Duration
+	HostServer string
+	Token      string
+	Timeout    time.Duration
 }
 
-func NewConnection(server string, token string, timeout time.Duration) *Connection {
-	return &Connection{Server: server, Token: token, Timeout: timeout}
+func NewConnection(hostServer string, token string, timeout time.Duration) *Connection {
+	return &Connection{HostServer: hostServer, Token: token, Timeout: timeout}
+}
+
+func (connection *Connection) GetHostServer() string {
+	return connection.HostServer
 }
 
 func (connection *Connection) DoGet(route string) (<-chan []byte, <-chan error) {
@@ -32,7 +36,7 @@ func (connection *Connection) DoGet(route string) (<-chan []byte, <-chan error) 
 
 		client := &http.Client{Timeout: connection.Timeout}
 
-		url := fmt.Sprintf("%s/%s", strings.TrimRight(connection.Server, "/"), strings.TrimLeft(route, "/"))
+		url := fmt.Sprintf("%s/%s", strings.TrimRight(connection.HostServer, "/"), strings.TrimLeft(route, "/"))
 		req, err := http.NewRequest("GET", url, nil)
 		if err != nil {
 			chErr <- err
