+++ sonar/factory/factory.go
@@ -1,11 +1,11 @@
 package factory
 
 import (
+	"sonarci/net"
 	"sonarci/sonar"
 	"sonarci/sonar/rest/v8"
-	"time"
 )
 
-func CreateLatestSonarRestApi(server string, token string, timeout time.Duration) sonar.Api {
-	return v8.NewRestApi(server, token, timeout)
+func CreateSonarApi(connection net.Connection) sonar.Api {
+	return v8.NewRestApi(connection)
 }
