+++ sonar/rest/v6/restapi.go
@@ -1,21 +1,14 @@
 package v6
 
 import (
-	"net/url"
 	"sonarci/net"
-	connFactory "sonarci/net/factory"
-	"time"
+	"sonarci/sonar/rest/base"
 )
 
 type RestApi struct {
-	net.Connection
-	Server string
+	base.RestApi
 }
 
-func NewRestApi(server string, token string, timeout time.Duration) *RestApi {
-	return &RestApi{Connection: connFactory.CreateHttpConnection(server, token, timeout)}
-}
-
-func escapeValue(value string) string {
-	return url.QueryEscape(value)
+func NewRestApi(connection net.Connection) *RestApi {
+	return &RestApi{*base.NewRestApi(connection)}
 }
