+++ sonar/rest/v6/validatebranch.go
@@ -1,71 +1,7 @@
 package v6
 
-import (
-	"encoding/json"
-	"errors"
-	"fmt"
-	"strings"
-)
-
-const routeBranchDetails = "/dashboard?id=%s&branch=%s"
-
-func (restApi *RestApi) GetRouteForValidateBranch() string {
-	return "/api/measures/component?componentKey=%s&branch=%s&metricKeys=alert_status"
-}
+const routeBranchValidation = "/api/measures/component?componentKey=%s&branch=%s&metricKeys=alert_status"
 
 func (restApi *RestApi) ValidateBranch(project string, branch string) error {
-	chBuff, chErr := restApi.DoGet(fmt.Sprintf(restApi.GetRouteForValidateBranch(), escapeValue(project), escapeValue(branch)))
-	err := <-chErr
-	if err != nil {
-		return err
-	}
-
-	buff := <-chBuff
-	wrapper := &branchStatusWrapper{}
-	err = json.Unmarshal(buff, wrapper)
-	if err != nil {
-		return err
-	}
-
-	wrapper.checkInfo(project, branch)
-	return restApi.validateBranchStatus(wrapper.Component)
-}
-
-func (restApi *RestApi) validateBranchStatus(status branchStatus) error {
-	const statusError = "ERROR"
-	if len(status.Measures) < 1 {
-		return errors.New(fmt.Sprintf("Failure on validate quality gate results\nFor more detail, visit: %s",
-			strings.TrimRight(restApi.Server, "/")+fmt.Sprintf(routeBranchDetails, escapeValue(status.Project), escapeValue(status.Branch))))
-	}
-
-	isValid := strings.ToUpper(status.Measures[0].Value) != statusError
-	if !isValid {
-		return errors.New(fmt.Sprintf("Branch %s has not been passed on quality gate\nFor more detail, visit: %s", escapeValue(status.Branch),
-			strings.TrimRight(restApi.Server, "/")+fmt.Sprintf(routeBranchDetails, escapeValue(status.Project), escapeValue(status.Branch))))
-	}
-
-	return nil
-}
-
-type branchStatusWrapper struct {
-	Component branchStatus `json:"component"`
-}
-
-type branchStatus struct {
-	Measures []branchStatusMeasure `json:"measures"`
-	Branch   string                `json:"branch"`
-	Project  string                `json:"key"`
-}
-
-type branchStatusMeasure struct {
-	Value string `json:"value"`
-}
-
-func (wrp *branchStatusWrapper) checkInfo(project string, branch string) {
-	if wrp.Component.Branch == "" {
-		wrp.Component.Branch = branch
-	}
-	if wrp.Component.Project == "" {
-		wrp.Component.Project = project
-	}
+	return restApi.ValidateBranchInternal(routeBranchValidation, project, branch)
 }
