+++ sonar/rest/v7/restapi.go
@@ -1,14 +1,14 @@
 package v7
 
 import (
-	"sonarci/sonar/rest/v6"
-	"time"
+	"sonarci/net"
+	v6 "sonarci/sonar/rest/v6"
 )
 
 type RestApi struct {
 	v6.RestApi
 }
 
-func NewRestApi(server string, token string, timeout time.Duration) *RestApi {
-	return &RestApi{*v6.NewRestApi(server, token, timeout)}
+func NewRestApi(connection net.Connection) *RestApi {
+	return &RestApi{*v6.NewRestApi(connection)}
 }
