+++ sonar/rest/v7/validatepullrequest.go
@@ -1,5 +1,7 @@
 package v7
 
-func (restApi *RestApi) GetRouteForValidatePullRequest() string {
-	return "/api/measures/component?component=%s&pullRequest=%s&metricKeys=alert_status"
+const routePullRequestValidation = "/api/measures/component?component=%s&pullRequest=%s&metricKeys=alert_status"
+
+func (restApi *RestApi) ValidatePullRequest(project string, pullRequest string) error {
+	return restApi.ValidatePullRequestInternal(routePullRequestValidation, project, pullRequest)
 }
