+++ sonar/rest/v8/restapi.go
@@ -1,14 +1,14 @@
 package v8
 
 import (
+	"sonarci/net"
 	"sonarci/sonar/rest/v7"
-	"time"
 )
 
 type RestApi struct {
 	v7.RestApi
 }
 
-func NewRestApi(server string, token string, timeout time.Duration) *RestApi {
-	return &RestApi{*v7.NewRestApi(server, token, timeout)}
+func NewRestApi(connection net.Connection) *RestApi {
+	return &RestApi{*v7.NewRestApi(connection)}
 }
